Add options to skip stitching step if processing untiled video
Add options to skip the per-GOP tile encoding and stitching interleaving steps.
These are unnecessary if the video is not tiled and we read all frames, regardless
of whether they contain objects.

When these options are specified, the query plan will directly output frames
from the encoder operator to the save operator. We need to make sure to
read frames in-order in this case.

TODO: unify option parameters and options dictionary.