Add more configuration options
Make it possible to specify the default database path and catalog path.

Also fix a bad class hierarchy between the SemanticIndex's. Create a virtual
base class for SQLite semantic indexes that both the XY and WH versions
inherit from to avoid opening both types of databases for the WH version.