Move database creation out of constructor
The class hierarchy of SemanticIndex is not the most ideal, and due to the
inheritance we would end up creating an on-disk database when constructing
an in-memory SemanticIndex when the base class's constructor was called.

We still want to share the implementation of all of the queries, so just
move the database setup out of the constructor. Create a factory class to
be responsible for initializing the correct class and then calling `setup()`.